[CBO] Unfork kqp_opt_stat
