+++ ydb/core/kqp/opt/cbo/solver/kqp_opt_stat.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "kqp_opt_predicate_selectivity.h"
+
 #include <ydb/library/yql/dq/opt/dq_opt.h>
 
 #include <ydb/core/kqp/opt/cbo/cbo_optimizer_new.h>
@@ -9,8 +11,6 @@ namespace NKikimr::NKqp {
 
 using namespace NYql::NNodes;
 using NYql::TExprNode;
-
-enum class EInequalityPredicateType : ui8 { Less, LessOrEqual, Greater, GreaterOrEqual, Equal };
 
 void InferStatisticsForFlatMap(const TExprNode::TPtr& input, TKqpStatsStore* kqpStats);
 void InferStatisticsForFilter(const TExprNode::TPtr& input, TKqpStatsStore* kqpStats);
@@ -59,101 +59,6 @@ TOrderingInfo GetSortBaseSortingOrderingInfo(const TCoSortBase&, const TSimpleSh
 TOrderingInfo GetAggregationBaseShuffleOrderingInfo(const TCoAggregateBase&, const TSimpleSharedPtr<TOrderingsStateMachine>& shufflingsFSM, TTableAliasMap*);
 TVector<TJoinColumn> GetKeySelectorOrdering(const TCoLambda& keySelector);
 
-class TPredicateSelectivityComputer {
-public:
-    struct TColumnStatisticsUsedMembers {
-        struct TColumnStatisticsUsedMember {
-            enum _ : ui32 {
-                EEquality,
-                EInequality
-            };
-
-            TColumnStatisticsUsedMember(TCoMember member, ui32 predicateType)
-                : Member(std::move(member))
-                , PredicateType(predicateType)
-            {}
-
-            TCoMember Member;
-            ui32 PredicateType;
-        };
-
-        void AddEquality(const TCoMember& member) {
-            Data.emplace_back(std::move(member), TColumnStatisticsUsedMember::EEquality);
-        }
-
-        void AddInequality(const TCoMember& member) {
-            Data.emplace_back(std::move(member), TColumnStatisticsUsedMember::EInequality);
-        }
-
-        TVector<TColumnStatisticsUsedMember> Data{};
-    };
-
-public:
-    TPredicateSelectivityComputer(
-        std::shared_ptr<TOptimizerStatistics> stats,
-        bool collectColumnsStatUsedMembers = false,
-        bool collectMemberEqualities = false,
-        bool collectConstantMembers = false
-    )
-        : Stats(std::move(stats))
-        , CollectColumnsStatUsedMembers(collectColumnsStatUsedMembers)
-        , CollectMemberEqualities(collectMemberEqualities)
-        , CollectConstantMembers(collectConstantMembers)
-    {}
-
-    double Compute(const TExprBase& input);
-
-    TColumnStatisticsUsedMembers GetColumnStatsUsedMembers() {
-        Y_ENSURE(CollectColumnsStatUsedMembers);
-        return ColumnStatsUsedMembers;
-    }
-
-    TVector<std::pair<TCoMember, TCoMember>> GetMemberEqualities() {
-        return MemberEqualities;
-    }
-
-    TVector<TCoMember> GetConstantMembers() {
-        return ConstantMembers;
-    }
-
-protected:
-    double ComputeImpl(
-        const TExprBase& input,
-        bool underNot,
-        bool collectConstantMembers
-    );
-
-    double ComputeEqualitySelectivity(
-        const TExprBase& left,
-        const TExprBase& right,
-        bool collectConstantMembers
-    );
-
-    double ComputeInequalitySelectivity(
-        const TExprBase& left,
-        const TExprBase& right,
-        EInequalityPredicateType predicate,
-        bool collectConstantMembers
-    );
-
-    double ComputeComparisonSelectivity(
-        const TExprBase& left,
-        const TExprBase& right
-    );
-
-private:
-    std::shared_ptr<TOptimizerStatistics> Stats;
-
-    bool CollectColumnsStatUsedMembers = false;
-    TColumnStatisticsUsedMembers ColumnStatsUsedMembers{};
-
-    bool CollectMemberEqualities = false;
-    TVector<std::pair<TCoMember, TCoMember>> MemberEqualities{};
-
-    bool CollectConstantMembers = false;
-    TVector<TCoMember> ConstantMembers{};
-};
-
 bool NeedCalc(TExprBase node);
 // Returns true if the expression is already a fully-evaluated literal
 // (TCoDataCtor, TCoNothing, or TCoJust wrapping a literal) and
@@ -161,15 +66,5 @@ bool NeedCalc(TExprBase node);
 bool IsLiteralDataExpr(TExprBase node);
 bool IsConstantExpr(const TExprNode::TPtr& input, bool foldUdfs = true);
 bool IsConstantExprWithParams(const TExprNode::TPtr& input);
-
-TCardinalityHints::TCardinalityHint* FindCardHint(TVector<TString>& labels, TCardinalityHints& hints);
-TCardinalityHints::TCardinalityHint* FindBytesHint(TVector<TString>& labels, TCardinalityHints& hints);
-std::shared_ptr<TOptimizerStatistics> ApplyBytesHints(std::shared_ptr<TOptimizerStatistics>& inputStats,
-    TVector<TString>& labels,
-    TCardinalityHints hints);
-std::shared_ptr<TOptimizerStatistics> ApplyRowsHints(
-    std::shared_ptr<TOptimizerStatistics>& inputStats,
-    TVector<TString>& labels,
-    TCardinalityHints hints);
 
 } // namespace NKikimr::NKqp
